feat(receipt): Receipt Scanner tool + Parse-as-receipt mode in OCR
Extract shared OcrWorkbench from ImageOcr (behavior-preserving); editable
ReceiptFields with JSON/CSV/copy export; new image-receipt-scanner tool.
Parsed data memoized per OCR result so edits survive re-renders.
